fix UB with pop16
this is UB because pop8 has side effects in which it increases SP after each read. the C standard states that these 2 functions can happen in any order, so the left side could happen last, which is not what we want to happen as the values would then be reversed!